% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bae_Luck_2018_data}
\alias{Bae_Luck_2018_data}
\title{Data from a motion estimation task}
\format{
A data frame with 20480 rows and 8 variables:
\describe{
\item{subject_Num}{observer ID}
\item{trial_Num}{trial number}
\item{TargetDirection}{true motion direction}
\item{RespAngle}{reported motion direction}
\item{motionCoh}{motion coherence}
\item{Block}{block number}
\item{Session}{session number}
\item{err}{estimation error}
}
}
\source{
\url{https://osf.io/4m2kb/download}
}
\usage{
Bae_Luck_2018_data
}
\description{
A dataset with the motion estimation results from Bae & Luck (2018) available from https://osf.io/4m2kb/ (some variables are removed, see the link for the full dataset).
}
\references{
{
Bae, G.-Y., & Luck, S. J. (2018). Decoding motion direction using the topography of sustained ERPs and alpha oscillations. NeuroImage, 184(August 2018), 242–255. \doi{10.1016/J.NEUROIMAGE.2018.09.029}
}
}
\keyword{datasets}
