% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{circ_corr}
\alias{circ_corr}
\title{Circular correlation coefficient}
\usage{
circ_corr(a, b, ill_defined = FALSE, mu = NULL, na.rm = FALSE)
}
\arguments{
\item{a}{first variable}

\item{b}{second variable}

\item{ill_defined}{is one of the variables mean is not well-defined (e.g., it is uniformly distributed)?}

\item{mu}{fix the mean parameter of both vectors to a certain value}

\item{na.rm}{a logical value indicating whether NA values should be removed before the computation proceeds}
}
\value{
correlation coefficient
}
\description{
Computes a circular correlation coefficient as defined in Jammalamadaka & SenGupta (2001).
}
\examples{
set.seed(1)
x <- rnorm(1000)
y <- 0.5 * x + sqrt(0.75) * rnorm(1000)
circ_corr(x, y)
}
\references{
{
Jammalamadaka, S. R., & SenGupta, A. (2001). Topics in Circular Statistics. WORLD SCIENTIFIC. \doi{10.1142/4031}
}
}
