% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circhelp-package.R
\docType{package}
\name{circhelp-package}
\alias{circhelp}
\alias{circhelp-package}
\title{circhelp: Circular Analyses Helper Functions}
\description{
Light-weight functions for computing descriptive statistics in different circular spaces (e.g., 2pi, 180, or 360 degrees), to handle angle-dependent biases, pad circular data, and more. Specifically aimed for psychologists and neuroscientists analyzing circular data. Basic methods are based on Jammalamadaka and SenGupta (2001) \doi{10.1142/4031}, removal of cardinal biases is based on the approach introduced in van Bergen, Ma, Pratte, & Jehee (2015) \doi{10.1038/nn.4150} and Chetverikov and Jehee (2023) \doi{10.1038/s41467-023-43251-w}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://achetverikov.github.io/circhelp/index.html}
  \item \url{https://github.com/achetverikov/circhelp}
  \item Report bugs at \url{https://github.com/achetverikov/circhelp/issues}
}

}
\author{
\strong{Maintainer}: Andrey Chetverikov \email{andrey.chetverikov@uib.no} (\href{https://orcid.org/0000-0003-2767-6310}{ORCID})

Authors:
\itemize{
  \item Andrey Chetverikov \email{andrey.chetverikov@uib.no} (\href{https://orcid.org/0000-0003-2767-6310}{ORCID})
}

Other contributors:
\itemize{
  \item Eline Van Geert \email{eline.vangeert@kuleuven.be} (\href{https://orcid.org/0000-0002-7848-5998}{ORCID}) [contributor]
}

}
\keyword{internal}
