% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correct_angle_rad}
\alias{correct_angle_rad}
\title{Correct angle value}
\usage{
correct_angle_rad(x)
}
\arguments{
\item{x}{angle}
}
\value{
angle in [-pi, pi] space
}
\description{
Correct angle value
}
\examples{
correct_angle_rad(4*pi)

}
