% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_plots_of_biases}
\alias{make_plots_of_biases}
\title{Plots biases using the data from \code{remove_cardinal_biases}}
\usage{
make_plots_of_biases(data, poly_deg, sd_val)
}
\arguments{
\item{data}{data prepared by  \code{remove_cardinal_biases}}

\item{poly_deg}{the degree of polynomial}

\item{sd_val}{sd used to determine the outliers}
}
\value{
does not return anything
}
\description{
Plots biases using the data from \code{remove_cardinal_biases}
}
\keyword{internal}
