% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pascucci_et_al_2019_data}
\alias{Pascucci_et_al_2019_data}
\title{Data from an orientation estimation task}
\format{
A data frame with 4400 rows and 5 variables:
\describe{
\item{observer}{observer ID}
\item{orientation}{true orientation}
\item{reported}{reported orientation}
\item{rt}{response time}
\item{err}{estimation error}
}
}
\source{
\url{https://zenodo.org/record/2544946/files/Experiment2_rawdata.csv?download=1}
}
\usage{
Pascucci_et_al_2019_data
}
\description{
A dataset with the orientation estimation results from Experiment 2 in Pascucci et al. (2019) available from https://doi.org/10.5281/zenodo.2544946.
}
\references{
{
Pascucci, D., Mancuso, G., Santandrea, E., Libera, C. D., Plomp, G., & Chelazzi, L. (2019). Laws of concatenated perception: Vision goes for novelty, decisions for perseverance. PLoS Biology, 17(3). \doi{10.1371/journal.pbio.3000144}
}
}
\keyword{datasets}
