% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smoothed_circ_sd}
\alias{smoothed_circ_sd}
\title{Compute smoothed estimate of circular standard deviation}
\usage{
smoothed_circ_sd(
  dt,
  circ_space = 180,
  weights_sd = 10,
  xvar = "abs_td_dist",
  yvar = "err",
  by = c(),
  n_points = 91
)
}
\arguments{
\item{dt}{data.table with the data.}

\item{circ_space}{Circular space, which can be 180 or 360 (default: 180).}

\item{weights_sd}{Standard deviation of the Gaussian window to use across \code{xvar} (default: 10).}

\item{xvar}{X-axis variable, such as dissimilarity between items (default: "abs_td_dist").}

\item{yvar}{Y-axis variable, normally errors (default: "err").}

\item{by}{A vector of grouping variable names (default: an empty vector).}

\item{n_points}{The number of points along the x-axis at which to compute circular SD (default: 91).}
}
\value{
A data.table with the grouping variables, \code{dist} - the values of X-axis variable at which the circular SD is computed, and \code{circ_sd} - the smoothed circular standard deviation estimate at each point.
}
\description{
This function calculates smoothed estimates of circular standard deviation using weighted kernel density estimation. It works similarly to \code{density_asymmetry} by applying Gaussian weights across an x-axis variable to compute localized circular SD estimates.
}
\examples{

data(Pascucci_et_al_2019_data)
ex_data <- Pascucci_et_al_2019_data
ex_data[, err := angle_diff_180(reported, orientation)] # response errors
ex_data[, prev_ori := shift(orientation), by = observer] # orientation on previous trial

# determine the shift in orientations between trials
ex_data[, diff_in_ori := angle_diff_180(prev_ori, orientation)]
ex_data[, abs_diff_in_ori := abs(diff_in_ori)]

circ_sd_smooth <- smoothed_circ_sd(ex_data[!is.na(diff_in_ori)],
  circ_space = 180, weights_sd = 10, xvar = "abs_diff_in_ori",
  yvar = "err", by = c("observer")
)

library(ggplot2)
ggplot(circ_sd_smooth, aes(x = dist, y = circ_sd)) +
  geom_line(stat = "summary", fun = mean) +
  labs(y = "Circular SD, °", x = "Absolute orientation difference, °")

}
