% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{circ_mean_rad}
\alias{circ_mean_rad}
\alias{circ_mean_180}
\alias{circ_mean_360}
\title{Circular mean}
\usage{
circ_mean_rad(x)

circ_mean_180(x)

circ_mean_360(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
mean of values in the vector
}
\description{
Circular mean
}
\section{Functions}{
\itemize{
\item \code{circ_mean_rad}: circular mean in 2pi space

\item \code{circ_mean_180}: circular mean in 180° space (e.g., line orientation)

\item \code{circ_mean_360}: circular mean in 360° space
}}

\examples{
x <- runif(1000,-pi,pi)
mean(x)
circ_mean_rad(x)

}
