% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circhelp-package.R
\docType{package}
\name{circhelp-package}
\alias{circhelp}
\alias{circhelp-package}
\title{circhelp: Circular Analyses Helper Functions}
\description{
A small helper package for circular data analyses, specifically aimed for psychologists and neuroscientists dealing with circular data. Includes light-weight functions for computing descriptive statistics in different circular spaces (e.g., 2pi, 180, or 360 degrees), functions to handle angle-dependent biases, padding circular data, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://achetverikov.github.io/circhelp/index.html}
  \item \url{https://github.com/achetverikov/circhelp}
}

}
\author{
\strong{Maintainer}: Andrey Chetverikov \email{andrey.chetverikov@uib.no} (\href{https://orcid.org/0000-0003-2767-6310}{ORCID})

Other contributors:
\itemize{
  \item Eline Van Geert \email{eline.vangeert@kuleuven.be} (\href{https://orcid.org/0000-0002-7848-5998}{ORCID}) [contributor]
}

}
\keyword{internal}
