% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density_asymmetry}
\alias{density_asymmetry}
\title{Compute asymmetry in weighted probability density}
\usage{
density_asymmetry(
  dt,
  circ_space = 180,
  weights_sd = 10,
  kernel_bw = NULL,
  xvar = "abs_td_dist",
  yvar = "bias_to_distr_corr",
  by = c(),
  n = 181,
  average = T,
  return_full_density = F
)
}
\arguments{
\item{dt}{data.table with the data}

\item{circ_space}{circular space (180 or 360)}

\item{weights_sd}{standard deviation of the Gaussian windiw to the use across xvar}

\item{kernel_bw}{bandwidth for the kernel density estimator across yvar}

\item{xvar}{x-axis variable (normally, dissimilarity)}

\item{yvar}{y-axis variable (normally, errors)}

\item{by}{a vector of grouping variables names}

\item{n}{the number of steps for the x-axis variable at which the density is computed}

\item{average}{average the asymmetry for each x-value (default: T)}

\item{return_full_density}{returns the full data.table with density computed at each point (default: F)}
}
\value{
data.table with several variables
}
\description{
Compute asymmetry in weighted probability density
}
\examples{

library(data.table)
data(Pascucci_et_al_2019_data)
ex_data <- Pascucci_et_al_2019_data
ex_data[, err := angle_diff_180(reported, orientation)] # response errors
ex_data[, prev_ori := shift(orientation), by = observer] # orientation on previous trial

# determine the shift in orientations between trials
ex_data[, diff_in_ori := angle_diff_180(prev_ori, orientation)]
ex_data[, abs_diff_in_ori := abs(diff_in_ori)]
ex_data[, err_rel_to_prev_targ := ifelse(diff_in_ori < 0, -err, err)]

err_dens <- density_asymmetry(ex_data[!is.na(err_rel_to_prev_targ)],
  circ_space = 180, weights_sd = 10, xvar = "abs_diff_in_ori",
  yvar = "err_rel_to_prev_targ", by = c("observer")
)

ggplot(err_dens, aes(x = dist, y = delta)) +
  geom_line(stat = "summary", fun = mean) +
  labs(y = "Asymmetry in error probability density, \%", x = "Absolute orientation difference, °")

}
