% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density_asymmetry}
\alias{density_asymmetry}
\title{Compute asymmetry in weighted probability density}
\usage{
density_asymmetry(
  dt,
  circ_space = 180,
  weights_sd = 10,
  kernel_bw = NULL,
  xvar = "abs_td_dist",
  yvar = "bias_to_distr_corr",
  by = c(),
  n = 181
)
}
\arguments{
\item{dt}{data.table with the data}

\item{circ_space}{circular space (180 or 360)}

\item{weights_sd}{standard deviation of the Gaussian windiw to the use across xvar}

\item{kernel_bw}{bandwidth for the kernel density estimator across yvar}

\item{xvar}{x-axis variable (normally, dissimilarity)}

\item{yvar}{y-axis variable (normally, errors)}

\item{by}{a vector of grouping variables names}

\item{n}{the number of steps for the x-axis variable at which the density is computed}
}
\value{
data.table with several variables
}
\description{
Compute asymmetry in weighted probability density
}
