% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_boundary_preds}
\alias{get_boundary_preds}
\title{Get polynomial predictions for values at the boundaries}
\usage{
get_boundary_preds(
  group_i,
  group_label,
  data,
  dist_to_centers_mat,
  space,
  reassign_range,
  poly_deg,
  angle_diff_fun,
  weights = NULL
)
}
\arguments{
\item{group_i}{integer group (bin) id}

\item{group_label}{group (bin) label}

\item{data}{dataset}

\item{dist_to_centers_mat}{precomputed distances to bin centers}

\item{space}{see \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}}

\item{reassign_range}{see \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}}

\item{poly_deg}{see \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}}

\item{angle_diff_fun}{a function to compute difference between angles}
}
\value{
a data.table with predicted values
}
\description{
A helper function for \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}.
}
\keyword{internal}
