% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_boundary_preds}
\alias{get_boundary_preds}
\title{Get polynomial predictions}
\usage{
get_boundary_preds(
  group,
  data,
  space,
  reassign_range,
  gam_ctrl,
  poly_deg,
  angle_diff_fun
)
}
\arguments{
\item{group}{group (bin) id}

\item{data}{dataset}

\item{space}{see \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}}

\item{reassign_range}{see \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}}

\item{gam_ctrl}{control object for gam models}

\item{poly_deg}{see \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}}

\item{angle_diff_fun}{a function to compute difference between angles}
}
\value{

}
\description{
A helper function for \code{\link[=remove_cardinal_biases]{remove_cardinal_biases()}}.
}
