% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.circ_loess}
\alias{predict.circ_loess}
\title{Compute predictions for circular LOESS}
\usage{
\method{predict}{circ_loess}(model, newdata, ...)
}
\arguments{
\item{model}{a circular LOESS object}

\item{newdata}{a data.frame with a variable x on which the predictions are computed}

\item{...}{other arguments (ignored)}
}
\value{
a data.frame with predictions
}
\description{
Compute predictions for circular LOESS
}
\keyword{internal}
